+++ vulture/__init__.py
@@ -1,4 +1,5 @@
-from vulture.core import __version__, Vulture
+from vulture.core import Vulture
+from vulture.version import __version__
 
 assert __version__
 assert Vulture